task 2 corrected

+++ 0x13-more_singly_linked_lists/1-listint_len.c
@@ -7,7 +7,7 @@
 size_t listint_len(const listint_t *h)
 {
 	const listint_t *traverse_ptr;
-	int node_count;
+	size_t node_count;
 
 	if (h == NULL)
 		return (1);
